Changed function read_Particles.
The function now checks first if particle is in the right domain. If not, the particle is simply deleted.
If yes, the index of the right cell in the subdomain is calculated (attention: this must be an inner cell).
To achieve this, the position of the particle must be corrected for the calculation only. This is done by
subtracting ic_lower_global and adding ic_start (modulo right cell_length factors) to the positions.