Wrote first step of comm1().
Function goes over all border cells (given that the border has a width of 1), checks the neighbouring
processes and calls functions comm1_send and comm1_recv.
comm1_send sends the particle-information of the current cell via Isend to the destination process.
comm1_recv receives the particle-information for the current cell via Recv and puts the particles in the particles_list of the current cell.