Change update_X_in.
Changed border_handling of function. The function now compares the particle positions to the global parameter of world (W_LC.s.L[dim]).
The positions of the particles are modified (case leaving) but the destination cell is calculated with a copy of corrected positions
which depends on the very subdomain.
The border handling for the periodic case will be done a second time in update_X before the particles of the world_particle-vector
are distributed over the cells (and border cells).
The periodic border-conditions are in fact handled by sending the particles to the right subdomain.